First code example with most elements integrated
For now, the code reads DMX input synchronously and sets channels 0 and
15 to their position based on channel 1.

Channels 2 and 3 define the min and max (respectively) duty cycle values
for `setPWM` since at least the servos we tested on seem to not match
the datasheet.

The input value for all DMX channels is output on the OLED display.