Changes PWM output connectors
The outputs will be broken out into a JST connector (with ground) in
blocks of 4 that can then be brought to the connectors to be used by the
end-user. This is a more flexible design, as we can have the break-out
board placed in a better place in the enclosure and also permits power
to be delivered directly to the breakout board.